working with curl...powershell script may be sketchy...checking...

+++ script.h
@@ -3,5 +3,5 @@ extern void _actr_log(const char* value, int length);
 extern void _actr_fillStyle(const char* value, int length);
 extern void _actr_strokeStyle(const char* value, int length);
 
-extern void _actr_fillRect(float x, float y, float w, float h);
-extern void _actr_strokeRect(float x, float y, float w, float h);
+extern void _actr_fillRect(int x, int y, int w, int h);
+extern void _actr_strokeRect(int x, int y, int w, int h);
